refactor the prompt utils into seperate files
