+++ vscode/src/autoedits/prompt/long-prompt-experimental.ts
@@ -6,15 +6,17 @@ import * as constants from './constants'
 import {
     getContextItemMappingWithTokenLimit,
     getContextItemsForIdentifier,
-    getCurrentFileLongSuggestionPrompt,
-    getLintErrorsPrompt,
     getPromptForTheContextSource,
     getPromptWithNewline,
-    getRecentEditsPrompt,
-    getRecentlyViewedSnippetsPrompt,
     joinPromptsWithNewlineSeparator,
-    groupConsecutiveRecentEditsItemsFromSameFile
-} from './prompt-utils'
+} from './prompt-utils/common'
+import { getCurrentFileLongSuggestionPrompt } from './prompt-utils/current-file'
+import { getLintErrorsPrompt } from './prompt-utils/lint'
+import {
+    getRecentEditsPrompt,
+    groupConsecutiveRecentEditsItemsFromSameFile,
+} from './prompt-utils/recent-edits'
+import { getRecentlyViewedSnippetsPrompt } from './prompt-utils/recent-view'
 
 export class LongTermPromptStrategy extends AutoeditsUserPromptStrategy {
     getUserPrompt({ context, tokenBudget, codeToReplaceData, document }: UserPromptArgs): PromptString {
