--- vscode/src/autoedits/prompt/prompt-utils.ts
@@ -1,583 +1,0 @@
-import * as vscode from 'vscode'
-import { groupConsecutiveItemsByPredicate } from '../../completions/context/retrievers/recent-user-actions/recent-edits-diff-helpers/utils'
-import {
-    type AutoEditsTokenLimit,
-    type AutocompleteContextSnippet,
-    type CodeToReplaceData,
-    type DocumentContext,
-    PromptString,
-    ps,
-    tokensToChars,
-} from '@sourcegraph/cody-shared'
-
-import { getTextFromNotebookCells } from '../../completions/context/retrievers/recent-user-actions/notebook-utils'
-import {
-    getActiveNotebookUri,
-    getCellIndexInActiveNotebookEditor,
-    getNotebookCells,
-} from '../../completions/context/retrievers/recent-user-actions/notebook-utils'
-import { RetrieverIdentifier } from '../../completions/context/utils'
-import {
-    getPrefixWithCharLimit,
-    getSuffixWithCharLimit,
-} from '../../completions/get-current-doc-context'
-import { autoeditsOutputChannelLogger } from '../output-channel-logger'
-import { clip, splitLinesKeepEnds } from '../utils'
-
-import * as constants from './constants'
-
-export { type CodeToReplaceData } from '@sourcegraph/cody-shared'
-
-export interface CurrentFilePromptOptions {
-    docContext: DocumentContext
-    document: vscode.TextDocument
-    position: vscode.Position
-    tokenBudget: Pick<
-        AutoEditsTokenLimit,
-        | 'codeToRewritePrefixLines'
-        | 'codeToRewriteSuffixLines'
-        | 'maxPrefixLinesInArea'
-        | 'maxSuffixLinesInArea'
-        | 'prefixTokens'
-        | 'suffixTokens'
-    >
-}
-
-export function getCompletionsPromptWithSystemPrompt(
-    systemPrompt: PromptString,
-    userPrompt: PromptString
-): PromptString {
-    // The models are offline fine-tuned on this prompt. It is important to keep it consistent.
-    return ps`${systemPrompt}\n\nUser: ${userPrompt}\n\nAssistant:`
-}
-
-export function getPromptWithNewline(prompt: PromptString): PromptString {
-    if (prompt.length === 0) {
-        return ps``
-    }
-    return ps`${prompt}\n`
-}
-
-export function getPromptForTheContextSource(
-    contextItems: AutocompleteContextSnippet[],
-    instructionPrompt: PromptString,
-    callback: (contextItems: AutocompleteContextSnippet[]) => PromptString
-): PromptString {
-    const prompt = callback(contextItems)
-    if (contextItems.length === 0 || prompt.length === 0) {
-        return ps``
-    }
-    return ps`${instructionPrompt}\n${prompt}`
-}
-
-export function getCurrentFileLongSuggestionPrompt({
-    document,
-    codeToReplaceDataRaw,
-}: {
-    document: vscode.TextDocument
-    codeToReplaceDataRaw: CodeToReplaceData
-    includeCursor?: boolean
-}): PromptString {
-    const filePath = getCurrentFilePath(document)
-    const codeToReplaceData = PromptString.fromAutoEditCodeToReplaceData(
-        codeToReplaceDataRaw,
-        document.uri
-    )
-
-    const prefix = ps`${codeToReplaceData.prefixBeforeArea}${codeToReplaceData.prefixInArea}`
-    const codeToRewrite = ps`${codeToReplaceData.codeToRewritePrefix}${constants.LONG_SUGGESTION_USER_CURSOR_MARKER}${codeToReplaceData.codeToRewriteSuffix}`
-    const suffix = ps`${codeToReplaceData.suffixInArea}${codeToReplaceData.suffixAfterArea}`
-
-    const areaPrompt = joinPromptsWithNewlineSeparator([
-        trimNewLineCharIfExists(prefix),
-        constants.LONG_SUGGESTION_EDITABLE_REGION_START_MARKER,
-        trimNewLineCharIfExists(codeToRewrite),
-        constants.LONG_SUGGESTION_EDITABLE_REGION_END_MARKER,
-        trimNewLineCharIfExists(suffix),
-    ])
-
-    const fileWithMarkerPrompt = getCurrentFileContextPromptWithPath(
-        filePath,
-        joinPromptsWithNewlineSeparator([constants.FILE_TAG_OPEN, areaPrompt, constants.FILE_TAG_CLOSE])
-    )
-    return fileWithMarkerPrompt
-}
-
-function trimNewLineCharIfExists(prompt: PromptString): PromptString {
-    if (prompt.toString().endsWith('\n')) {
-        return prompt.slice(0, -1)
-    }
-    return prompt
-}
-
-export function getCurrentFilePromptComponents({
-    document,
-    codeToReplaceDataRaw,
-    includeCursor,
-}: {
-    document: vscode.TextDocument
-    codeToReplaceDataRaw: CodeToReplaceData
-    includeCursor?: boolean
-}): {
-    fileWithMarkerPrompt: PromptString
-    areaPrompt: PromptString
-} {
-    const filePath = getCurrentFilePath(document)
-    const codeToReplaceData = PromptString.fromAutoEditCodeToReplaceData(
-        codeToReplaceDataRaw,
-        document.uri
-    )
-
-    const fileWithMarker = joinPromptsWithNewlineSeparator([
-        codeToReplaceData.prefixBeforeArea,
-        constants.AREA_FOR_CODE_MARKER,
-        codeToReplaceData.suffixAfterArea,
-    ])
-
-    const fileWithMarkerPrompt = getCurrentFileContextPromptWithPath(
-        filePath,
-        joinPromptsWithNewlineSeparator([
-            constants.FILE_TAG_OPEN,
-            fileWithMarker,
-            constants.FILE_TAG_CLOSE,
-        ])
-    )
-
-    const codeToRewrite = includeCursor
-        ? ps`${codeToReplaceData.codeToRewritePrefix}<CURSOR_IS_HERE>${codeToReplaceData.codeToRewriteSuffix}`
-        : codeToReplaceData.codeToRewrite
-
-    const areaPrompt = joinPromptsWithNewlineSeparator([
-        constants.AREA_FOR_CODE_MARKER_OPEN,
-        codeToReplaceData.prefixInArea,
-        constants.CODE_TO_REWRITE_TAG_OPEN,
-        codeToRewrite,
-        constants.CODE_TO_REWRITE_TAG_CLOSE,
-        codeToReplaceData.suffixInArea,
-        constants.AREA_FOR_CODE_MARKER_CLOSE,
-    ])
-
-    return {
-        fileWithMarkerPrompt,
-        areaPrompt,
-    }
-}
-
-export function getCodeToReplaceData(options: CurrentFilePromptOptions): CodeToReplaceData {
-    const {
-        position,
-        document,
-        docContext,
-        tokenBudget: {
-            codeToRewritePrefixLines,
-            codeToRewriteSuffixLines,
-            maxPrefixLinesInArea,
-            maxSuffixLinesInArea,
-            prefixTokens,
-            suffixTokens,
-        },
-    } = options
-
-    const numContextLines = splitLinesKeepEnds(docContext.prefix + docContext.suffix).length
-    const numPrefixLines = splitLinesKeepEnds(docContext.prefix).length
-    const adjustment = position.character !== 0 ? 1 : 0
-
-    const minLine = clip(position.line - numPrefixLines + adjustment, 0, document.lineCount - 1)
-    const maxLine = clip(minLine + numContextLines - 1, 0, document.lineCount - 1)
-
-    const codeToRewriteStart = clip(position.line - codeToRewritePrefixLines, minLine, maxLine)
-    const codeToRewriteEnd = clip(position.line + codeToRewriteSuffixLines, minLine, maxLine)
-    const areaStart = clip(
-        position.line - maxPrefixLinesInArea - codeToRewritePrefixLines,
-        minLine,
-        maxLine
-    )
-    const areaEnd = clip(
-        position.line + maxSuffixLinesInArea + codeToRewriteSuffixLines,
-        minLine,
-        maxLine
-    )
-
-    // Helper function to create position from line number
-    const positionAtLineStart = (line: number) => new vscode.Position(line, 0)
-    const positionAtLineEnd = (line: number) => document.lineAt(line).rangeIncludingLineBreak.end
-
-    // Create ranges for different sections
-    const ranges = {
-        codeToRewrite: new vscode.Range(
-            positionAtLineStart(codeToRewriteStart),
-            positionAtLineEnd(codeToRewriteEnd)
-        ),
-        codeToRewritePrefix: new vscode.Range(
-            positionAtLineStart(codeToRewriteStart),
-            new vscode.Position(position.line, position.character)
-        ),
-        codeToRewriteSuffix: new vscode.Range(
-            new vscode.Position(position.line, position.character),
-            positionAtLineEnd(codeToRewriteEnd)
-        ),
-        prefixInArea: new vscode.Range(
-            positionAtLineStart(areaStart),
-            positionAtLineStart(codeToRewriteStart)
-        ),
-        suffixInArea: new vscode.Range(positionAtLineEnd(codeToRewriteEnd), positionAtLineEnd(areaEnd)),
-        prefixBeforeArea: new vscode.Range(positionAtLineStart(minLine), positionAtLineStart(areaStart)),
-        suffixAfterArea: new vscode.Range(positionAtLineEnd(areaEnd), positionAtLineEnd(maxLine)),
-    }
-
-    const remainingPrefixChars = Math.max(0, tokensToChars(prefixTokens) - docContext.prefix.length)
-    const remainingSuffixChars = Math.max(0, tokensToChars(suffixTokens) - docContext.suffix.length)
-
-    const { prefixBeforeArea, suffixAfterArea } = getUpdatedCurrentFilePrefixAndSuffixOutsideArea(
-        document,
-        ranges.prefixBeforeArea,
-        ranges.suffixAfterArea,
-        remainingPrefixChars,
-        remainingSuffixChars
-    )
-
-    return {
-        codeToRewrite: document.getText(ranges.codeToRewrite),
-        codeToRewritePrefix: document.getText(ranges.codeToRewritePrefix),
-        codeToRewriteSuffix: document.getText(ranges.codeToRewriteSuffix),
-        prefixInArea: document.getText(ranges.prefixInArea),
-        suffixInArea: document.getText(ranges.suffixInArea),
-        prefixBeforeArea: prefixBeforeArea.toString(),
-        suffixAfterArea: suffixAfterArea.toString(),
-        range: ranges.codeToRewrite,
-    }
-}
-
-export function getCurrentFilePath(document: vscode.TextDocument): PromptString {
-    const uri =
-        document.uri.scheme === 'vscode-notebook-cell'
-            ? getActiveNotebookUri() ?? document.uri
-            : document.uri
-    return PromptString.fromDisplayPath(uri)
-}
-
-function getUpdatedCurrentFilePrefixAndSuffixOutsideArea(
-    document: vscode.TextDocument,
-    rangePrefixBeforeArea: vscode.Range,
-    rangeSuffixAfterArea: vscode.Range,
-    remainingPrefixChars: number,
-    remainingSuffixChars: number
-): {
-    prefixBeforeArea: PromptString
-    suffixAfterArea: PromptString
-} {
-    const { prefixBeforeAreaForNotebook, suffixAfterAreaForNotebook } =
-        getPrefixAndSuffixForAreaForNotebook(document, remainingPrefixChars, remainingSuffixChars)
-
-    const prefixBeforeArea = ps`${prefixBeforeAreaForNotebook}${PromptString.fromDocumentText(
-        document,
-        rangePrefixBeforeArea
-    )}`
-
-    const suffixAfterArea = ps`${PromptString.fromDocumentText(
-        document,
-        rangeSuffixAfterArea
-    )}${suffixAfterAreaForNotebook}`
-
-    return {
-        prefixBeforeArea,
-        suffixAfterArea,
-    }
-}
-
-function getPrefixAndSuffixForAreaForNotebook(
-    document: vscode.TextDocument,
-    remainingPrefixChars: number,
-    remainingSuffixChars: number
-): {
-    prefixBeforeAreaForNotebook: PromptString
-    suffixAfterAreaForNotebook: PromptString
-} {
-    const currentCellIndex = getCellIndexInActiveNotebookEditor(document)
-    if (currentCellIndex === -1) {
-        return {
-            prefixBeforeAreaForNotebook: ps``,
-            suffixAfterAreaForNotebook: ps``,
-        }
-    }
-    const activeNotebook = vscode.window.activeNotebookEditor?.notebook!
-    const notebookCells = getNotebookCells(activeNotebook)
-    const cellsBeforeCurrentCell = notebookCells.slice(0, currentCellIndex)
-    const cellsAfterCurrentCell = notebookCells.slice(currentCellIndex + 1)
-    const beforeContent = getTextFromNotebookCells(activeNotebook, cellsBeforeCurrentCell)
-    const afterContent = getTextFromNotebookCells(activeNotebook, cellsAfterCurrentCell)
-
-    const beforeContentList = beforeContent.split('\n')
-    const afterContentList = afterContent.split('\n')
-
-    return {
-        prefixBeforeAreaForNotebook: ps`${getPrefixWithCharLimit(
-            beforeContentList,
-            remainingPrefixChars
-        )}\n`,
-        suffixAfterAreaForNotebook: ps`\n${getSuffixWithCharLimit(
-            afterContentList,
-            remainingSuffixChars
-        )}`,
-    }
-}
-
-export function getLintErrorsPrompt(contextItems: AutocompleteContextSnippet[]): PromptString {
-    const lintErrors = getContextItemsForIdentifier(
-        contextItems,
-        RetrieverIdentifier.DiagnosticsRetriever
-    )
-    if (lintErrors.length === 0) {
-        return ps``
-    }
-
-    // Create a mapping of URI to AutocompleteContextSnippet[]
-    const uriToSnippetsMap = new Map<string, AutocompleteContextSnippet[]>()
-    for (const item of lintErrors) {
-        const uriString = item.uri.toString()
-        if (!uriToSnippetsMap.has(uriString)) {
-            uriToSnippetsMap.set(uriString, [])
-        }
-        uriToSnippetsMap.get(uriString)!.push(item)
-    }
-
-    // Combine snippets for each URI
-    const combinedPrompts: PromptString[] = []
-    for (const [uriString, snippets] of uriToSnippetsMap) {
-        const uri = vscode.Uri.parse(uriString)
-        const snippetContents = snippets.map(
-            item => PromptString.fromAutocompleteContextSnippet(item).content
-        )
-        const combinedContent = joinPromptsWithNewlineSeparator(snippetContents, ps`\n\n`)
-        const promptWithPath = getContextPromptWithPath(
-            PromptString.fromDisplayPath(uri),
-            combinedContent
-        )
-        combinedPrompts.push(promptWithPath)
-    }
-
-    const lintErrorsPrompt = joinPromptsWithNewlineSeparator(combinedPrompts, ps`\n\n`)
-    return joinPromptsWithNewlineSeparator([
-        constants.LINT_ERRORS_TAG_OPEN,
-        lintErrorsPrompt,
-        constants.LINT_ERRORS_TAG_CLOSE,
-    ])
-}
-
-export function getRecentCopyPrompt(contextItems: AutocompleteContextSnippet[]): PromptString {
-    const recentCopy = getContextItemsForIdentifier(
-        contextItems,
-        RetrieverIdentifier.RecentCopyRetriever
-    )
-    if (recentCopy.length === 0) {
-        return ps``
-    }
-    const recentCopyPrompts = recentCopy.map(item =>
-        getContextPromptWithPath(
-            PromptString.fromDisplayPath(item.uri),
-            PromptString.fromAutocompleteContextSnippet(item).content
-        )
-    )
-    const recentCopyPrompt = joinPromptsWithNewlineSeparator(recentCopyPrompts, ps`\n\n`)
-    return joinPromptsWithNewlineSeparator([
-        constants.RECENT_COPY_TAG_OPEN,
-        recentCopyPrompt,
-        constants.RECENT_COPY_TAG_CLOSE,
-    ])
-}
-
-export function getRecentEditsPrompt(contextItems: AutocompleteContextSnippet[]): PromptString {
-    const recentEdits = getContextItemsForIdentifier(
-        contextItems,
-        RetrieverIdentifier.RecentEditsRetriever
-    )
-    recentEdits.reverse()
-    if (recentEdits.length === 0) {
-        return ps``
-    }
-    const recentEditsPrompts = recentEdits.map(item =>
-        getRecentEditsContextPromptWithPath(
-            PromptString.fromDisplayPath(item.uri),
-            PromptString.fromAutocompleteContextSnippet(item).content
-        )
-    )
-    const recentEditsPrompt = joinPromptsWithNewlineSeparator(recentEditsPrompts)
-    return joinPromptsWithNewlineSeparator([
-        constants.RECENT_EDITS_TAG_OPEN,
-        recentEditsPrompt,
-        constants.RECENT_EDITS_TAG_CLOSE,
-    ])
-}
-
-export function getRecentlyViewedSnippetsPrompt(
-    contextItems: AutocompleteContextSnippet[]
-): PromptString {
-    const recentViewedSnippets = getContextItemsForIdentifier(
-        contextItems,
-        RetrieverIdentifier.RecentViewPortRetriever
-    )
-    recentViewedSnippets.reverse()
-    if (recentViewedSnippets.length === 0) {
-        return ps``
-    }
-    const recentViewedSnippetPrompts = recentViewedSnippets.map(item =>
-        joinPromptsWithNewlineSeparator([
-            constants.SNIPPET_TAG_OPEN,
-            getContextPromptWithPath(
-                PromptString.fromDisplayPath(item.uri),
-                PromptString.fromAutocompleteContextSnippet(item).content
-            ),
-            constants.SNIPPET_TAG_CLOSE,
-        ])
-    )
-
-    const snippetsPrompt = joinPromptsWithNewlineSeparator(recentViewedSnippetPrompts)
-    return joinPromptsWithNewlineSeparator([
-        constants.RECENT_SNIPPET_VIEWS_TAG_OPEN,
-        snippetsPrompt,
-        constants.RECENT_SNIPPET_VIEWS_TAG_CLOSE,
-    ])
-}
-
-export function getJaccardSimilarityPrompt(contextItems: AutocompleteContextSnippet[]): PromptString {
-    const jaccardSimilarity = getContextItemsForIdentifier(
-        contextItems,
-        RetrieverIdentifier.JaccardSimilarityRetriever
-    )
-    if (jaccardSimilarity.length === 0) {
-        return ps``
-    }
-    const jaccardSimilarityPrompts = jaccardSimilarity.map(item =>
-        joinPromptsWithNewlineSeparator([
-            constants.SNIPPET_TAG_OPEN,
-            getContextPromptWithPath(
-                PromptString.fromDisplayPath(item.uri),
-                PromptString.fromAutocompleteContextSnippet(item).content
-            ),
-            constants.SNIPPET_TAG_CLOSE,
-        ])
-    )
-
-    const snippetsPrompt = joinPromptsWithNewlineSeparator(jaccardSimilarityPrompts)
-    return joinPromptsWithNewlineSeparator([
-        constants.EXTRACTED_CODE_SNIPPETS_TAG_OPEN,
-        snippetsPrompt,
-        constants.EXTRACTED_CODE_SNIPPETS_TAG_CLOSE,
-    ])
-}
-
-//  Helper functions
-export function getContextItemMappingWithTokenLimit(
-    contextItems: AutocompleteContextSnippet[],
-    contextTokenLimitMapping: Record<string, number>,
-    contextNumItemsLimitMapping: Record<string, number>
-): Map<RetrieverIdentifier, AutocompleteContextSnippet[]> {
-    const contextItemMapping = new Map<RetrieverIdentifier, AutocompleteContextSnippet[]>()
-    // Group items by identifier
-    for (const item of contextItems) {
-        const identifier = item.identifier as RetrieverIdentifier
-        if (!contextItemMapping.has(identifier)) {
-            contextItemMapping.set(identifier, [])
-        }
-        contextItemMapping.get(identifier)!.push(item)
-    }
-    // Apply token limits
-    for (const [identifier, items] of contextItemMapping) {
-        const tokenLimit =
-            identifier in contextTokenLimitMapping ? contextTokenLimitMapping[identifier] : undefined
-        const numItemsLimit =
-            identifier in contextNumItemsLimitMapping ? contextNumItemsLimitMapping[identifier] : undefined
-
-        if (tokenLimit !== undefined) {
-            contextItemMapping.set(identifier, getContextItemsInTokenBudget(items, tokenLimit, numItemsLimit))
-        } else {
-            autoeditsOutputChannelLogger.logError(
-                'getContextItemMappingWithTokenLimit',
-                `No token limit for ${identifier}`
-            )
-            contextItemMapping.set(identifier, [])
-        }
-    }
-    return contextItemMapping
-}
-
-export function getContextItemsInTokenBudget(
-    contextItems: AutocompleteContextSnippet[],
-    tokenBudget: number,
-    numItemsLimit?: number
-): AutocompleteContextSnippet[] {
-    const autocompleteItemsWithBudget: AutocompleteContextSnippet[] = []
-    let currentCharsCount = 0
-    const charsBudget = tokensToChars(tokenBudget)
-
-    for (let i = 0; i < contextItems.length; i++) {
-        const item = contextItems[i]
-        if (currentCharsCount + item.content.length > charsBudget) {
-            continue
-        }
-        currentCharsCount += item.content.length
-        autocompleteItemsWithBudget.push(item)
-        if (numItemsLimit && autocompleteItemsWithBudget.length >= numItemsLimit) {
-            break
-        }
-    }
-    return autocompleteItemsWithBudget
-}
-
-export function groupConsecutiveRecentEditsItemsFromSameFile(contextItems: AutocompleteContextSnippet[]): AutocompleteContextSnippet[] {
-    if (contextItems.length === 0) {
-        return []
-    }
-    // Group consecutive items by file name
-    const groupedContextItems = groupConsecutiveItemsByPredicate(
-        contextItems,
-        (a, b) => a.uri.toString() === b.uri.toString()
-    )
-    const combinedContextItems: AutocompleteContextSnippet[] = []
-    for (const group of groupedContextItems) {
-        const combinedItem = {
-            ...group[0],
-            // The group content is from the latest to the oldest item.
-            // We need to reverse the order of the content to get diff from old to new.
-            content: group
-                .map(item => item.content)
-                .reverse()
-                .join('\nthen\n'),
-        }
-        combinedContextItems.push(combinedItem)
-    }
-    return combinedContextItems
-}
-
-export function getContextItemsForIdentifier(
-    contextItems: AutocompleteContextSnippet[],
-    identifier: string
-): AutocompleteContextSnippet[] {
-    return contextItems.filter(item => item.identifier === identifier)
-}
-
-export function getContextPromptWithPath(filePath: PromptString, content: PromptString): PromptString {
-    return ps`(\`${filePath}\`)\n\n${content}`
-}
-
-function getCurrentFileContextPromptWithPath(
-    filePath: PromptString,
-    content: PromptString
-): PromptString {
-    return ps`(\`${filePath}\`)\n${content}`
-}
-
-export function getRecentEditsContextPromptWithPath(
-    filePath: PromptString,
-    content: PromptString
-): PromptString {
-    return ps`${filePath}\n${content}`
-}
-
-export function joinPromptsWithNewlineSeparator(
-    prompts: PromptString[],
-    separator = ps`\n`
-): PromptString {
-    const validPrompts = prompts.filter(args => args.length > 0)
-    return PromptString.join(validPrompts, separator)
-}
