+++ vscode/src/autoedits/shrink-prediction.ts
@@ -1,6 +1,6 @@
 import { getNewLineChar, lines } from '../completions/text-processing'
 
-import type { CodeToReplaceData } from './prompt/prompt-utils'
+import type { CodeToReplaceData } from '@sourcegraph/cody-shared'
 
 /**
  * Shrinks the prediction by removing overlapping lines with the suffix.
